⚡️ Implement Optimistic UI for instant updates
Optimistic UI allows to react to user input changes even before they are confirmed in the bac kend. To implement the optimistic interface, we separated the update code into a separate function and called it in the onMutate event, which is triggered the moment the user clicks the save button. We also use the onError event to deal with errors in the request, reverting the cache to the original data if a problem occurs. The optimistic interface is a powerful technique for improving the user experience in situations where the probability of error is low, it instantly updates information before they are confirmed in the back end.