✨ Implement status operations and integrate with API
1. All the files inside the API folder.
2. order-table-row.tsx

Moved the function responsible for gathering the cache data and updating the order status into a separate function (updateOrderStatusOnCache).

Created a mutateAsync function for each API call file, responsible for triggering the API call and, upon success (onSuccess), invoking the function responsible for updating the cache.

Implemented these functions on each button to constantly check the order status and retrieve the current button dynamically.

Created a custom size for the button because the xs value was altering the table size when the text was small. Now, with a custom size with a fixed width, this issue has been resolved.