🥅 Error handle and 404 page
Using errorElement is not the best way to deal with page not found, as it is also used for internal errors. Instead, we split it into two elements: one for the page not found and one for internal errors.

This code shows how to create a custom error page and how to use the useRouteError hook to get information about the error that occurred.

In production, it is not recommended to show the error to the user, but rather to send it to an error management platform.